requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate
Flask-Script
Flask-Cors
psycopg2-binary